requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI/ML
google-generativeai==0.3.2
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.3.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP Clients
aiohttp==3.9.1
httpx==0.26.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
